// web/src/types/chat/conversation_message.ts
export const enum ConversationMessageType {
	/** 文本消息 */
	Text = "text",

	/** 富文本消息 */
	RichText = "rich_text",

	/** markdown 消息 */
	Markdown = "markdown",

	/** AI 搜索 */
	AggregateAISearchCard = "aggregate_ai_search_card",

	/** 图片消息 */
	Image = "image",

	/** 视频消息 */
	Video = "video",

	/** 文件消息 */
	Files = "files",

	/** 语音消息 */
	Voice = "voice",

	/** 魔法搜索卡片消息 */
	MagicSearchCard = "magic_search_card",

	/** AI 文生图 */
	AiImage = "ai_image_card",

	/** 原图转高清 */
	HDImage = "image_convert_high_card",

	/** 录音纪要 */
	RecordingSummary = "recording_summary",
	
}

/**
 * 携带推理内容和流式消息状态
 */
export interface WithReasoningContentAndStreamOptions {
	reasoning_content?: string
	stream_options?: {
		/** 流式消息状态 */
		status: StreamStatus
		/** 流式消息 ID */
		stream: boolean
	}
}

/**
 * 会话消息
 */
export interface ConversationMessageBase extends SeqMessageBase {
	/** 应用消息 ID */
	app_message_id: string
	/** 发送者 ID */
	sender_id: string
	/** 发送时间 */
	send_time: number
	/** 消息状态 */
	status: ConversationMessageStatus
	/** 未读消息数 */
	unread_count: number
	/** 话题 ID */
	topic_id?: string
	/** 是否撤回 */
	revoked?: boolean
	/** 本地临时数据存储 */
	temp_custom_data?: Record<string, unknown>
}

/**
 * 消息附件
 */
export type ConversationMessageAttachment = {
	file_id: string
	file_type?: number
	file_extension?: string
	file_size?: number
	file_name?: string
}

/**
 * 会话消息指令
 */
export type ConversationMessageInstruct = {
	value: string
	instruction: Exclude<QuickInstruction, SystemInstruct>
}

/**
 * 消息携带指令
 */
export type WithConversationMessageInstruct = {
	instructs?: ConversationMessageInstruct[] | undefined
}
// ...
export interface RichTextConversationMessage extends ConversationMessageBase {
	type: ConversationMessageType.RichText
	/** 消息内容 */
	rich_text?: {
		content: string
		attachments?: ConversationMessageAttachment[]
	} & WithConversationMessageInstruct
}

/**
 * 文本消息
 */
export interface TextConversationMessage extends ConversationMessageBase {
	type: ConversationMessageType.Text
	/** 消息内容 */
	text?: {
		content: string
		attachments?: ConversationMessageAttachment[]
	} & WithReasoningContentAndStreamOptions &
		WithConversationMessageInstruct
}

export type RecordSummaryOriginContent = {
	duration: string
	speaker: string
	text: string
	start_time: string
	end_time: string
}[]

/**
 * 录音纪要消息
 */
export interface RecordSummaryConversationMessage extends ConversationMessageBase {
	type: ConversationMessageType.RecordingSummary
	/** 消息内容 */
	recording_summary?: {
		status?: RecordSummaryStatus
		text?: string // 单次翻译的结果
		recording_blob?: string // 翻译的数据源
		origin_content?: RecordSummaryOriginContent // 原文内容
		attachments?: ConversationMessageAttachment[] // 音频文件
		full_text?: string // 当前完整的翻译结果
		ai_result?: string // 智能总结后的结果
		duration?: string // 总时长
		title?: string // 智能总结标题
		audio_link?: string // 录音链接
		is_recognize?: boolean // 是否进行识别
	}
}

/**
 * markdown 消息
 */
export interface MarkdownConversationMessage extends ConversationMessageBase {
	type: ConversationMessageType.Markdown
	/** 消息内容 */
	markdown?: {
		content: string
		attachments?: ConversationMessageAttachment[]
	} & WithReasoningContentAndStreamOptions &
		WithConversationMessageInstruct
}
// ...
export const enum RecordSummaryStatus {
	/** 开始录音 */
	Start = 1,
	/** 录音中 */
	Doing = 2,
	/** 结束录音 */
	End = 3,
	/** 录音总结中 */
	Summarizing = 4,
	/** 录音总结完成 */
	Summarized = 5,
}

/**
 * AI 搜索 - 事件
 */
export interface AggregateAISearchCardEvent {
	name: string
	time: string
	description: string
}

export const enum AggregateAISearchCardDeepLevel {
	/** 简单搜索 */
	Simple = 1,
	/** 深度搜索 */
	Deep = 2,
}

/**
 * AI 搜索内容
 */
interface AggregateAISearchCardContentFromService extends WithReasoningContentAndStreamOptions {
	/** 父级 ID */
	parent_id: string
	/** 当前 ID */
	id: string
	/** 类型 */
	type: AggregateAISearchCardDataType
	/** 回答 */
	llm_response: string | undefined
	/** 关联问题 */
	/** FIXME: 后端代码 返回类型改为 AssociateQuestion, 后续可移除 string */
	associate_questions: Record<string, string | AssociateQuestion> | undefined
	/** 思维导图 */
	mind_map: AggregateAISearchCardMindMapChildren | string | undefined
	/** 搜索结果 */
	search: AggregateAISearchCardSearch[] | undefined
	/** 关键词 */
	search_keywords: string[] | undefined
	/** 事件 */
	event: AggregateAISearchCardEvent[] | undefined
	/** 搜索深度, 1 - 简单搜索 2 - 深度搜索 */
	search_deep_level: AggregateAISearchCardDeepLevel | undefined
	/** PPT */
	ppt: string | undefined
	/** 总字数 */
	total_words: number | undefined
	/** 检索到的页面总数 */
	match_count: number | undefined
	/** 阅读的页面总数 */
	page_count: number | undefined
}

export type AggregateAISearchCardContent = Pick<
	AggregateAISearchCardContentFromService,
	| "mind_map"
	| "llm_response"
	| "event"
	| "ppt"
	| "search_deep_level"
	| "reasoning_content"
	| "stream_options"
> & {
	search: Record<string, AggregateAISearchCardSearch[]>
	associate_questions: Record<string, AssociateQuestion>
	finish: boolean
	error: boolean
}

/**
 * AI 搜索消息
 */
export interface AggregateAISearchCardConversationMessage<FromService extends boolean = true>
	extends ConversationMessageBase {
	type: ConversationMessageType.AggregateAISearchCard
	aggregate_ai_search_card?: FromService extends true
		? AggregateAISearchCardContentFromService
		: AggregateAISearchCardContent
}

/**
 * 文件消息
 */
export interface FileConversationMessage extends ConversationMessageBase {
	type: ConversationMessageType.Files
	/** 消息内容 */
	files?: {
		attachments: ConversationMessageAttachment[]
	} & WithConversationMessageInstruct
}

/**
 * 图片消息
 */
export interface ImageConversationMessage extends ConversationMessageBase {
	type: ConversationMessageType.Image
	/** 消息内容 */
	image?: {
		file_id: string
	} & WithConversationMessageInstruct
}

/**
 * 视频消息
 */
export interface VideoConversationMessage extends ConversationMessageBase {
	type: ConversationMessageType.Video
	/** 消息内容 */
	video?: {
		file_id: string
	} & WithConversationMessageInstruct
}

/**
 * 语音消息
 */
export interface VoiceConversationMessage extends ConversationMessageBase {
	type: ConversationMessageType.Voice
	/** 消息内容 */
	voice?: {
		file_id: string
	} & WithConversationMessageInstruct
}
// ...
export interface MagicSearchCardContent {
	/** 搜索结果 */
	search: WikiSearchItem[]
	/** 回答 */
	llm_response: string
	/** 相关问题 */
	related_questions: string[]
}

/**
 * 魔法搜索卡片消息
 */
export interface MagicSearchCardConversationMessage extends ConversationMessageBase {
	type: ConversationMessageType.MagicSearchCard
	magic_search_card?: MagicSearchCardContent
}

/**
 * 撤回消息
 */
export interface RevokeMessage extends ConversationMessageBase {
	type: ControlEventMessageType.RevokeMessage
	revoke_message: {
		/** 撤回消息 ID */
		refer_message_id: string
	}
}

/**
 * 群创建消息
 */
export interface GroupCreateMessage extends ConversationMessageBase {
	type: ControlEventMessageType.GroupCreate
	/** 未读数 */
	unread_count: number
	/** 发送时间 */
	send_time: number
	/** 状态 */
	status: ConversationMessageStatus
	group_create: {
		/** 操作人ID */
		operate_user_id: string
		/** 群ID */
		group_id: string
		/** 用户ID列表 */
		user_ids: string[]
		/** 会话ID */
		conversation_id: string
		/** 群名称 */
		group_name: string
		/** 群主ID */
		group_owner_id: string
	}
}

/**
 * 群解散消息
 */
export interface GroupDisbandMessage extends ConversationMessageBase {
	type: ControlEventMessageType.GroupDisband
	/** 未读数 */
	unread_count: number
	/** 发送时间 */
	send_time: number
	/** 状态 */
	status: ConversationMessageStatus
}

/**
 * 群新增成员消息
 */
export interface GroupAddMemberMessage extends ConversationMessageBase {
	type: ControlEventMessageType.GroupAddMember
	/** 未读数 */
	unread_count: number
	/** 发送时间 */
	send_time: number
	/** 状态 */
	status: ConversationMessageStatus
	/** 群新增成员 */
	group_users_add: {
		/** 操作人ID */
		operate_user_id: string
		/** 群ID */
		group_id: string
		/** 用户ID列表 */
		user_ids: string[]
		/** 会话ID */
		conversation_id: string
	}
}

/**
 * 群人员退群
 */
export interface GroupUsersRemoveMessage extends ConversationMessageBase {
	type: ControlEventMessageType.GroupUsersRemove
	/** 未读数 */
	unread_count: number
	/** 发送时间 */
	send_time: number
	/** 状态 */
	status: ConversationMessageStatus
	group_users_remove: {
		/** 操作人ID */
		operate_user_id: string
		/** 群ID */
		group_id: string
		/** 用户ID列表 */
		user_ids: string[]
		/** 会话ID */
		conversation_id: string
	}
}

/**
 * 群更新
 */
export interface GroupUpdateMessage extends ConversationMessageBase {
	type: ControlEventMessageType.GroupUpdate
	/** 未读数 */
	unread_count: number
	/** 发送时间 */
	send_time: number
	/** 状态 */
	status: ConversationMessageStatus
	group_update: {
		/** 操作人ID */
		operate_user_id: string
		/** 群ID */
		group_id: string
		/** 会话ID */
		conversation_id: string
		/** 群名称 */
		group_name: string
		/** 群头像 */
		group_avatar: string
	}
}
// ...
export interface AIImagesContent {
	// 生成状态
	type: AIImagesDataType
	// 图片列表
	items: AIImagesContentItem[]
	// 文本描述
	text: string
	// 引用文件
	refer_file_id: string
	// 文本描述
	refer_text: string
	// 错误信息
	error_message?: string
	// 比例
	radio?: string
}

/**
 * AI 文生图
 */
export interface AIImagesMessage extends ConversationMessageBase {
	type: ConversationMessageType.AiImage
	/** 消息内容 */
	ai_image_card?: AIImagesContent
}

export enum HDImageDataType {
	/** 开始生成 */
	StartGenerate = 1,
	/** 生成完成 */
	GenerateComplete = 2,
	/** 异常终止 */
	Error = 3,
}

export interface HDImageContent {
	// 生成状态
	type: HDImageDataType
	// 原图文件id
	origin_file_id: string
	// 高清图文件id
	new_file_id: string
	// 文本描述
	refer_text: string
	// 错误信息
	error_message?: string
	// 比例
	radio?: string
}

/**
 * 原图转高清
 */
export interface HDImageMessage extends ConversationMessageBase {
	type: ConversationMessageType.HDImage
	/** 消息内容 */
	image_convert_high_card?: HDImageContent
}
// ...
export type ConversationMessage =
	| TextConversationMessage
	| RichTextConversationMessage
	| MarkdownConversationMessage
	| MagicSearchCardConversationMessage
	| FileConversationMessage
	| ImageConversationMessage
	| VideoConversationMessage
	| VoiceConversationMessage
	| RevokeMessage
	| GroupAddMemberMessage
	| GroupCreateMessage
	| GroupDisbandMessage
	| GroupUsersRemoveMessage
	| GroupUpdateMessage
	| AggregateAISearchCardConversationMessage<false>
	| AIImagesMessage
	| RecordSummaryConversationMessage
	| HDImageMessage
// ...
export const enum ConversationMessageStatus {
	/** 未读 */
	Unread = "unread",
	/** 已读 */
	Read = "read",
	/** 已看 */
	Seen = "seen",
	/** 已撤回 */
	Revoked = "revoked",
}